adds feeds to the new router
